+++ NEWS.md
@@ -2,7 +2,13 @@
 title: "NEWS for the emmeans package"
 ---
 
-## emmeans 1.9.0-90xxxx
+## emmeans 1.10.0-90xxxxxx
+  * With `gls` or `lme` models, `mode = "satterthwaite"`
+    and `mode =  "appx-satterthwaite"` failed when model was fitted with no
+    explicit `data` argument (#465)
+   
+
+## emmeans 1.10.0
   * Restored `tau` argument (now optional) for rq models (#458)
   * Fixed issue where a multivariate factor having numeric levels may
     mismatch a level in `at` even when apparently valid (#458)